Adding Ceph deployment support to the Makefile
This change introduces the Ceph related targets with the goal of
deploying an external cluster that lives within OCP but is reacheable
from the outside.
With this patch users should be able to run:

'make ceph' to:

1. prepare the host (a ceph-osd-losetup for a dummy osd)
2. deploy the rook operator
3. copy the resulting CephCluster CR with the same pattern used for the
other operators

'make ceph_deploy' uses kustomize to apply the CephCluster CR.
'make ceph_cleanup' is the target used to tear down the cluster, remove
the operator and cleanup the host removing the ceph data files created
in /var/lib, as well as removing the dummy OSD related files.

Notes:

* HostNetworking is used to expose the cluster outside
* The cluster is not deployed in the 'openstack' namespace, but
  it uses 'rook-ceph' by default (we can avoid mixing things)

Signed-off-by: Francesco Pantano <fpantano@redhat.com>